--- .cafehub/Dockerfile.bt
@@ -1,39 +1,0 @@
-FROM python:latest
-
-RUN apt-get update
-
-RUN apt-get install -y bluez bluetooth libmtdev-dev vim xauth
-
-RUN apt-get install -y openjdk-17-jdk
-
-ENV JDK_HOME=/usr/lib/jvm/java-17-openjdk-arm64
-
-RUN python -m pip install --upgrade pip
-
-RUN pip install virtualenv
-
-RUN touch /root/.Xauthority
-
-RUN python -m virtualenv kivy_venv
-
-RUN ln -s kivy_venv/bin/activate
-
-RUN pip3 install cython
-
-RUN pip3 install kivy
-
-RUN pip3 install oscpy
-
-RUN pip3 install bleak==0.17.0
-
-RUN pip3 install pydantic
-
-RUN pip3 install websocket-server
-
-RUN pip3 install jnius
-
-RUN pip3 install pytest
-
-WORKDIR /root
-
-ENTRYPOINT bash
