--- .cafehub/Dockerfile.cafehub
@@ -1,13 +1,0 @@
-FROM bt
-
-RUN apt-get install -y git
-
-WORKDIR /root
-
-RUN git clone https://github.com/rheasman/cafehub/
-
-RUN echo "#!/bin/sh\n\nservice dbus start\nservice bluetooth start\n\nbash" > entry.sh
-
-RUN chmod +x entry.sh
-
-CMD /root/entry.sh
