+++ src/App.tsx
@@ -7,9 +7,8 @@ import WideView from './components/WideView'
 import usePreventNavigatingAwayEffect from '$/hooks/usePreventNavigatingAwayEffect'
 import NarrowView from '$/components/NarrowView'
 import { Container } from 'toasterhea'
-import { Layer } from './consts'
+import { Layer } from './types'
 import { useUiStore } from './stores/ui'
-import Debug from './components/ui/Debug'
 import { useAutoConnectEffect } from './stores/data'
 
 const App = () => {
@@ -35,7 +34,6 @@ const App = () => {
             >
                 <WideView tw="hidden lg:block" />
                 <NarrowView tw="lg:hidden" />
-                {/* <Debug /> */}
             </div>
             <Container id={Layer.Drawer} />
         </>
