+++ src/components/drawers/ProfilesDrawer.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 import Drawer, { DrawerProps } from '$/components/drawers/Drawer'
 import ListItem from '$/components/ListItem'
 import tw from 'twin.macro'
-import { profiles } from '$/consts'
-import { useCafeHubStore } from '$/stores/ch'
+import { profiles } from '$/types'
+import { useDataStore } from '$/stores/data'
 
 interface ProfilesDrawerProps extends Pick<DrawerProps, 'onReject'> {
     onResolve?: () => void
 }
 
 export default function ProfilesDrawer({ onReject, onResolve }: ProfilesDrawerProps) {
-    const { profile: currentProfile, setProfile } = useCafeHubStore()
+    const { profile: currentProfile, setProfile } = useDataStore()
 
     return (
         <Drawer
