--- src/hooks/useMinorState.ts
@@ -1,9 +1,0 @@
-import { MinorState } from '$/consts'
-import { Property } from '$/types'
-import { usePropertyValue } from '$/stores/ch'
-
-export function useMinorState(): MinorState {
-    return usePropertyValue(Property.MinorState, {
-        defaultValue: MinorState.NoState,
-    })
-}
